Moved restaurant controller to /http/controllers
